test of node.js bindings in electron application success

+++ app/native/src/Logger.cpp
@@ -28,7 +28,7 @@ namespace Comp {
 
   void Logger::setLogLevel(LogLevel level)
   {
-    _level = level;
+    _level = (int)level;
   }
 
   string Logger::printTime()
